Retire wk-connect (#7032)
* remove isConnector property for datastores

* snapshots, test db

* delete datastore in migration. add changelog

* changelog + migration guide

* adapt schema number

* adapt schema version numbers

* remove Neuroglancer and BossDB dataset tabs

* fix TS types

* updated docs to remove BossDB and wkconnect

* remove wkconnect from docker file

* removed isReadOnlyDataset from dataset settings

* remove wkconnect datastore from dummy test data

* update snapshots

* fix test db schema

* refresh snapshots

* Update CHANGELOG.unreleased.md


* Update MIGRATIONS.unreleased.md


* Update docs/export.md


* Update docs/index.md


* refresh snapshots

---------
