+++ CHANGELOG.unreleased.md
@@ -17,5 +17,6 @@ For upgrade instructions, please check the [migration guide](MIGRATIONS.released
 ### Fixed
 
 ### Removed
+- Support for [webknososs-connect](https://github.com/scalableminds/webknossos-connect) data store servers has been removed. Use the "Add Remote Dataset" functionality instead. [#7031](https://github.com/scalableminds/webknossos/pull/7031)
 
 ### Breaking Changes
