+++ app/controllers/DataStoreController.scala
@@ -30,7 +30,6 @@ class DataStoreController @Inject()(dataStoreDAO: DataStoreDAO,
       (__ \ 'publicUrl).read[String] and
       (__ \ 'key).read[String] and
       (__ \ 'isScratch).readNullable[Boolean] and
-      (__ \ 'isConnector).readNullable[Boolean] and
       (__ \ 'allowsUpload).readNullable[Boolean])(DataStore.fromForm _)
 
   private val dataStorePublicReads: Reads[DataStore] =
@@ -38,7 +37,6 @@ class DataStoreController @Inject()(dataStoreDAO: DataStoreDAO,
       (__ \ 'url).read[String] and
       (__ \ 'publicUrl).read[String] and
       (__ \ 'isScratch).readNullable[Boolean] and
-      (__ \ 'isConnector).readNullable[Boolean] and
       (__ \ 'allowsUpload).readNullable[Boolean])(DataStore.fromUpdateForm _)
   @ApiOperation(value = "List all available datastores", nickname = "datastoreList")
   @ApiResponses(
