+++ conf/evolutions/102-no-more-wkconnect.sql
@@ -1,0 +1,15 @@
+START TRANSACTION;
+
+do $$ begin ASSERT (select schemaVersion from webknossos.releaseInformation) = 101, 'Previous schema version mismatch'; end; $$ LANGUAGE plpgsql;
+
+DROP VIEW webknossos.dataStores_;
+
+DELETE FROM webknossos.dataStores WHERE isConnector;
+
+ALTER TABLE webknossos.dataStores DROP COLUMN isConnector;
+
+CREATE VIEW webknossos.dataStores_ AS SELECT * FROM webknossos.dataStores WHERE NOT isDeleted;
+
+UPDATE webknossos.releaseInformation SET schemaVersion = 102;
+
+COMMIT TRANSACTION;
