+++ docs/faq.md
@@ -31,9 +31,9 @@ WEBKNOSSOS does not allow you to run custom machine learning models on your data
 
 ## How can I use my dataset with WEBKNOSSOS?
 
-WEBKNOSSOS supports [WKW (optimized), KNOSSOS cubes](./datasets.md), and image stacks (converted on upload). You can also connect to [Neuroglancer Precomputed dataset hosted in the Google Cloud and to data hosted by a BossDB](https://github.com/scalableminds/webknossos-connect).
+WEBKNOSSOS supports [WKW (optimized), KNOSSOS cubes](./datasets.md), and image stacks (converted on upload). You can also connect to Neuroglancer Precomputed, N5, and Zarr datasets hosted in the cloud (Google Cloud Storage, AWS S3).
 
-Smaller datasets (up to multiple GB) can be uploaded directly through the web interface. For larger datasets, we recommend converting them to the standard WKW format using the [WEBKNOSSOS Cuber](https://docs.webknossos.org/wkcuber/index.html) CLI tool and uploading it via the [WEBKNOSSOS python package](https://docs.webknossos.org/webknossos-py/examples/upload_image_data.html).
+Smaller datasets (up to multiple GB) can be uploaded directly through the web interface. For larger datasets, we recommend converting them to the standard WKW format using the [WEBKNOSSOS Cuber](https://docs.webknossos.org/wkcuber/index.html) CLI tool and uploading it via the [WEBKNOSSOS Python package](https://docs.webknossos.org/webknossos-py/examples/upload_image_data.html).
 
 ## Can I host the WEBKNOSSOS data in my own compute cluster (on-premise installation)?
