+++ frontend/javascripts/admin/admin_rest_api.ts
@@ -52,7 +52,6 @@ import type {
   ExperienceDomainList,
   ServerTracing,
   TracingType,
-  WkConnectDatasetConfig,
   ServerEditableMapping,
   APICompoundType,
   ZarrPrivateLink,
@@ -1537,19 +1536,6 @@ export function cancelDatasetUpload(
   );
 }
 
-export function addWkConnectDataset(
-  datastoreHost: string,
-  datasetConfig: WkConnectDatasetConfig,
-): Promise<void> {
-  return doWithToken((token) =>
-    Request.sendJSONReceiveJSON(`/data/datasets?token=${token}`, {
-      data: datasetConfig,
-      host: datastoreHost,
-      method: "POST",
-    }),
-  );
-}
-
 type ExplorationResult = {
   dataSource: DatasourceConfiguration | undefined;
   report: string;
@@ -1847,7 +1833,7 @@ export async function getPublication(id: string): Promise<APIPublication> {
 }
 
 // #### Datastores
-export async function getDatastores(): Promise<Array<APIDataStore>> {
+export async function getDatastores(): Promise<APIDataStore[]> {
   const datastores = await Request.receiveJSON("/api/datastores");
   assertResponseLimit(datastores);
   return datastores;
