+++ frontend/javascripts/admin/onboarding.tsx
@@ -417,7 +417,7 @@ class OnboardingView extends React.PureComponent<Props, State> {
   }
 
   async fetchDatastores() {
-    const datastores = (await getDatastores()).filter((ds) => !ds.isConnector);
+    const datastores = await getDatastores();
     this.setState({
       datastores,
     });
