+++ frontend/javascripts/test/fixtures/skeletontracing_server_objects.ts
@@ -181,7 +181,6 @@ export const annotation: APIAnnotation = {
     name: "localhost",
     url: "http://localhost:9000",
     isScratch: false,
-    isConnector: false,
     allowsUpload: true,
   },
   tracingStore: {
