+++ frontend/javascripts/test/fixtures/tasktracing_server_objects.ts
@@ -126,7 +126,6 @@ export const annotation: APIAnnotation = {
     name: "localhost",
     url: "http://localhost:9000",
     isScratch: false,
-    isConnector: false,
     allowsUpload: true,
   },
   tracingStore: {
