+++ frontend/javascripts/test/fixtures/volumetracing_server_objects.ts
@@ -95,7 +95,6 @@ export const annotation: APIAnnotation = {
     name: "localhost",
     url: "http://localhost:9000",
     isScratch: false,
-    isConnector: false,
     allowsUpload: true,
   },
   tracingStore: {
