+++ frontend/javascripts/test/snapshots/public-test/test-bundle/test/backend-snapshot-tests/annotations.e2e.js.md
@@ -18,7 +18,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: '2012-06-28_Cortex',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
@@ -137,7 +136,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: '2012-06-28_Cortex',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
@@ -256,7 +254,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: '2012-06-28_Cortex',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
@@ -427,7 +424,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: '2012-06-28_Cortex',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
@@ -598,7 +594,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: '2012-06-28_Cortex',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
@@ -722,7 +717,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: '2012-06-28_Cortex',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
@@ -846,7 +840,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: '2012-06-28_Cortex',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
@@ -965,7 +958,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: 'confocal-multi_knossos',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
