+++ frontend/javascripts/test/snapshots/public-test/test-bundle/test/backend-snapshot-tests/datasets.e2e.js.md
@@ -253,7 +253,6 @@ Generated by [AVA](https://avajs.dev).
         },
         dataStore: {
           allowsUpload: true,
-          isConnector: false,
           isScratch: false,
           name: 'localhost',
           url: 'http://localhost:9000',
@@ -391,7 +390,6 @@ Generated by [AVA](https://avajs.dev).
         },
         dataStore: {
           allowsUpload: true,
-          isConnector: false,
           isScratch: false,
           name: 'localhost',
           url: 'http://localhost:9000',
