+++ frontend/javascripts/test/snapshots/public-test/test-bundle/test/backend-snapshot-tests/misc.e2e.js.md
@@ -8,15 +8,7 @@ Generated by [AVA](https://avajs.dev).
 
     [
       {
-        allowsUpload: false,
-        isConnector: true,
-        isScratch: false,
-        name: 'connect',
-        url: 'http://localhost:8000',
-      },
-      {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
