+++ frontend/javascripts/test/snapshots/public-test/test-bundle/test/backend-snapshot-tests/tasks.e2e.js.md
@@ -412,7 +412,6 @@ Generated by [AVA](https://avajs.dev).
         dataSetName: '2012-06-28_Cortex',
         dataStore: {
           allowsUpload: true,
-          isConnector: false,
           isScratch: false,
           name: 'localhost',
           url: 'http://localhost:9000',
@@ -697,7 +696,6 @@ Generated by [AVA](https://avajs.dev).
       dataSetName: 'confocal-multi_knossos',
       dataStore: {
         allowsUpload: true,
-        isConnector: false,
         isScratch: false,
         name: 'localhost',
         url: 'http://localhost:9000',
