+++ test/db/dataStores.csv
@@ -1,3 +1,2 @@
-name,url,publicUrl,key,isScratch,isDeleted,isConnector,allowsUpload,onlyAllowedOrganization,reportUsedStorageEnabled
-'localhost','http://localhost:9000','http://localhost:9000','something-secure',f,f,f,t,,f
-'connect','http://localhost:8000','http://localhost:8000','secret-key',f,f,t,f,,f
+name,url,publicUrl,key,isScratch,isDeleted,allowsUpload,onlyAllowedOrganization,reportUsedStorageEnabled
+'localhost','http://localhost:9000','http://localhost:9000','something-secure',f,f,t,,f
